mnma-upload/requirements.txt
fastapi==0.111.0
uvicorn[standard]
PyYAML
boto3
pydantic
nest_asyncio
asyncio==3.4.3
fastapi-cognito
python-multipart
tenacity
pymysql
numpy>=1.21.0
aiofiles>=23.0.0
bcrypt>=4.0.0
passlib>=1.7.4